fix: SectionOverrides interpreted as any
