+++ packages/core/src/components/sections/EmptyState/Overrides.tsx
@@ -1,23 +1,8 @@
-import type { PropsWithChildren } from 'react'
 import { EmptyState as UIEmptyState } from '@faststore/ui'
-import type { EmptyStateProps } from '@faststore/ui'
 
-import type {
-  SectionOverrideDefinition,
-  ComponentOverrideDefinition,
-} from 'src/typings/overrides'
 import { getSectionOverrides } from 'src/utils/overrides'
 import { override } from 'src/customizations/components/overrides/EmptyState'
-
-export type EmptyStateOverrideDefinition = SectionOverrideDefinition<
-  'EmptyState',
-  {
-    EmptyState: ComponentOverrideDefinition<
-      PropsWithChildren<EmptyStateProps>,
-      EmptyStateProps
-    >
-  }
->
+import type { EmptyStateOverrideDefinition } from 'src/typings/overrides'
 
 const { EmptyState } = getSectionOverrides(
   {
