+++ packages/core/src/components/sections/Navbar/Overrides.tsx
@@ -11,67 +11,10 @@ import {
   NavbarButtons as UINavbarButtons,
   IconButton as UIIconButton,
 } from '@faststore/ui'
-import type {
-  IconButtonProps,
-  NavbarButtonsProps,
-  NavbarHeaderProps,
-  NavbarLinksListProps,
-  NavbarLinksProps,
-  NavbarProps,
-  NavbarRowProps,
-  NavbarSliderContentProps,
-  NavbarSliderFooterProps,
-  NavbarSliderHeaderProps,
-  NavbarSliderProps,
-} from '@faststore/ui'
 
-import type {
-  ComponentOverrideDefinition,
-  SectionOverrideDefinition,
-} from 'src/typings/overrides'
 import { getSectionOverrides } from 'src/utils/overrides'
 import { override } from 'src/customizations/components/overrides/Navbar'
-
-export type NavbarOverrideDefinition = SectionOverrideDefinition<
-  'Navbar',
-  {
-    Navbar: ComponentOverrideDefinition<NavbarProps, NavbarProps>
-    NavbarLinks: ComponentOverrideDefinition<NavbarLinksProps, NavbarLinksProps>
-    NavbarLinksList: ComponentOverrideDefinition<
-      NavbarLinksListProps,
-      NavbarLinksListProps
-    >
-    NavbarSlider: ComponentOverrideDefinition<
-      NavbarSliderProps,
-      NavbarSliderProps
-    >
-    NavbarSliderHeader: ComponentOverrideDefinition<
-      NavbarSliderHeaderProps,
-      NavbarSliderHeaderProps
-    >
-    NavbarSliderContent: ComponentOverrideDefinition<
-      NavbarSliderContentProps,
-      NavbarSliderContentProps
-    >
-    NavbarSliderFooter: ComponentOverrideDefinition<
-      NavbarSliderFooterProps,
-      NavbarSliderFooterProps
-    >
-    NavbarHeader: ComponentOverrideDefinition<
-      NavbarHeaderProps,
-      NavbarHeaderProps
-    >
-    NavbarRow: ComponentOverrideDefinition<NavbarRowProps, NavbarRowProps>
-    NavbarButtons: ComponentOverrideDefinition<
-      NavbarButtonsProps,
-      NavbarButtonsProps
-    >
-    IconButton: ComponentOverrideDefinition<
-      IconButtonProps,
-      Omit<IconButtonProps, 'onClick'>
-    >
-  }
->
+import type { NavbarOverrideDefinition } from 'src/typings/overrides'
 
 const {
   Navbar,
