+++ packages/core/src/components/sections/ProductGallery/Overrides.tsx
@@ -6,13 +6,6 @@ import {
   Icon as UIIcon,
 } from '@faststore/ui'
 
-import type {
-  ButtonProps,
-  IconProps,
-  LinkButtonProps,
-  SkeletonProps,
-} from '@faststore/ui'
-
 import ProductCard from 'src/components/product/ProductCard'
 import FilterDesktop from 'src/components/search/Filter/FilterDesktop'
 
@@ -20,47 +13,9 @@ const FilterSlider = lazy(
   () => import('src/components/search/Filter/FilterSlider')
 )
 
-import type {
-  SectionOverrideDefinition,
-  ComponentOverrideDefinition,
-} from 'src/typings/overrides'
 import { getSectionOverrides } from 'src/utils/overrides'
 import { override } from 'src/customizations/components/overrides/ProductGallery'
-
-export type ProductGalleryOverrideDefinition = SectionOverrideDefinition<
-  'ProductGallery',
-  {
-    Button: ComponentOverrideDefinition<
-      ButtonProps,
-      Omit<ButtonProps, 'onClick'>
-    >
-    FilterIcon: ComponentOverrideDefinition<IconProps, IconProps>
-    PrevIcon: ComponentOverrideDefinition<IconProps, IconProps>
-    ResultsCountSkeleton: ComponentOverrideDefinition<
-      SkeletonProps,
-      Omit<SkeletonProps, 'loading'>
-    >
-    SortSkeleton: ComponentOverrideDefinition<
-      SkeletonProps,
-      Omit<SkeletonProps, 'loading'>
-    >
-    FilterButtonSkeleton: ComponentOverrideDefinition<
-      SkeletonProps,
-      Omit<SkeletonProps, 'loading'>
-    >
-    LinkButtonPrev: ComponentOverrideDefinition<
-      LinkButtonProps,
-      Omit<LinkButtonProps, 'onClick' | 'href'>
-    >
-    LinkButtonNext: ComponentOverrideDefinition<
-      LinkButtonProps,
-      Omit<LinkButtonProps, 'onClick' | 'href'>
-    >
-    __experimentalFilterDesktop: ComponentOverrideDefinition<any, any>
-    __experimentalFilterSlider: ComponentOverrideDefinition<any, any>
-    __experimentalProductCard: ComponentOverrideDefinition<any, any>
-  }
->
+import type { ProductGalleryOverrideDefinition } from 'src/typings/overrides'
 
 const {
   Button,
