+++ packages/core/src/components/sections/ProductShelf/Overrides.tsx
@@ -1,29 +1,11 @@
 import { ProductShelf as UIProductShelf } from '@faststore/ui'
-import type { ProductShelfProps } from '@faststore/ui'
 
 import ProductCard from 'src/components/product/ProductCard'
 import Carousel from 'src/components/ui/Carousel'
-import type {
-  ComponentOverrideDefinition,
-  SectionOverrideDefinition,
-} from 'src/typings/overrides'
+
 import { getSectionOverrides } from 'src/utils/overrides'
 import { override } from 'src/customizations/components/overrides/ProductShelf'
-
-export type ProductShelfOverrideDefinition = SectionOverrideDefinition<
-  'ProductShelf',
-  {
-    ProductShelf: ComponentOverrideDefinition<
-      ProductShelfProps,
-      ProductShelfProps
-    >
-    __experimentalCarousel: ComponentOverrideDefinition<any, any>
-    __experimentalProductCard: ComponentOverrideDefinition<
-      any,
-      Omit<any, 'key' | 'product' | 'index'>
-    >
-  }
->
+import type { ProductShelfOverrideDefinition } from 'src/typings/overrides'
 
 const { ProductShelf, __experimentalCarousel, __experimentalProductCard } =
   getSectionOverrides(
