+++ packages/core/src/typings/overrides.ts
@@ -1,68 +1,267 @@
-import type { AlertOverrideDefinition } from '../components/sections/Alert/Overrides'
-import type { BannerTextOverrideDefinition } from '../components/sections/BannerText/Overrides'
-import type { BreadcrumbOverrideDefinition } from '../components/sections/Breadcrumb/Overrides'
-import type { EmptyStateOverrideDefinition } from '../components/sections/EmptyState/Overrides'
-import type { HeroOverrideDefinition } from '../components/sections/Hero/Overrides'
-import type { NavbarOverrideDefinition } from '../components/sections/Navbar/Overrides'
-import type { NewsletterOverrideDefinition } from '../components/sections/Newsletter/Overrides'
-import type { ProductDetailsOverrideDefinition } from '../components/sections/ProductDetails/Overrides'
-import type { ProductGalleryOverrideDefinition } from '../components/sections/ProductGallery/Overrides'
-import type { ProductShelfOverrideDefinition } from '../components/sections/ProductShelf/Overrides'
-import type { RegionBarOverrideDefinition } from '../components/sections/RegionBar/Overrides'
+import type { PropsWithChildren } from 'react'
+import type {
+  AlertProps,
+  BannerTextContentProps,
+  BannerTextProps,
+  BreadcrumbProps,
+  ButtonProps,
+  DiscountBadgeProps,
+  EmptyStateProps,
+  HeroHeaderProps,
+  HeroImageProps,
+  HeroProps,
+  IconButtonProps,
+  IconProps,
+  ImageGalleryProps,
+  ImageZoomProps,
+  InputFieldProps,
+  LinkButtonProps,
+  NavbarButtonsProps,
+  NavbarHeaderProps,
+  NavbarLinksListProps,
+  NavbarLinksProps,
+  NavbarProps,
+  NavbarRowProps,
+  NavbarSliderContentProps,
+  NavbarSliderFooterProps,
+  NavbarSliderHeaderProps,
+  NavbarSliderProps,
+  PriceProps,
+  ProductShelfProps,
+  ProductTitleProps,
+  QuantitySelectorProps,
+  RegionBarProps,
+  ShippingSimulationProps,
+  SkeletonProps,
+  SkuSelectorProps,
+} from '@faststore/ui'
+
+import type {
+  ComponentOverrideDefinition,
+  SectionOverrideDefinition,
+} from './overrideDefinitionUtils'
 
 export type SectionOverride =
-  | NewsletterOverrideDefinition
-  | RegionBarOverrideDefinition
-  | HeroOverrideDefinition
   | AlertOverrideDefinition
   | BannerTextOverrideDefinition
   | BreadcrumbOverrideDefinition
   | EmptyStateOverrideDefinition
+  | HeroOverrideDefinition
   | NavbarOverrideDefinition
-  | ProductShelfOverrideDefinition
+  | NewsletterOverrideDefinition
   | ProductDetailsOverrideDefinition
   | ProductGalleryOverrideDefinition
+  | ProductShelfOverrideDefinition
+  | RegionBarOverrideDefinition
 
-export type GetSectionOverridesReturn<SO extends SectionOverride> = {
-  [Key in keyof SO['components']]: Merge<SO['components'][Key]>
-}
+/**
+ * Originally, these types were defined in their respective Overrides file
+ * For some reason, typescript wouldn't interpret SectionOverride correctly when
+ * defining it by importing each of the SectionOverrideDefinitions and would set it as 'any'
+ *
+ * For some reason, defining them in the same file as SectionOverride seems to fix the issue.
+ * Consider that before moving them elsewhere and test it on the starter as well.
+ */
+export type AlertOverrideDefinition = SectionOverrideDefinition<
+  'Alert',
+  {
+    Alert: ComponentOverrideDefinition<AlertProps, Omit<AlertProps, 'onClose'>>
+    Icon: ComponentOverrideDefinition<IconProps, IconProps>
+  }
+>
 
-export type DefaultSectionComponentsDefinitions<
-  T extends SectionOverrideDefinition
-> = Record<keyof T['components'], React.ComponentType>
+export type BannerTextOverrideDefinition = SectionOverrideDefinition<
+  'BannerText',
+  {
+    BannerText: ComponentOverrideDefinition<BannerTextProps, BannerTextProps>
+    BannerTextContent: ComponentOverrideDefinition<
+      BannerTextContentProps,
+      BannerTextContentProps
+    >
+  }
+>
 
-export type SectionOverrideDefinition<
-  SectionName extends string = string,
-  OC extends Record<string, ComponentOverrideDefinition<any, any>> = Record<
-    string,
-    ComponentOverrideDefinition<any, any>
-  >
-> = {
-  section: SectionName
-  components?: Partial<Prettify<OC>>
-}
+export type BreadcrumbOverrideDefinition = SectionOverrideDefinition<
+  'Breadcrumb',
+  {
+    Breadcrumb: ComponentOverrideDefinition<BreadcrumbProps, BreadcrumbProps>
+    Icon: ComponentOverrideDefinition<IconProps, IconProps>
+  }
+>
 
-export type ComponentOverrideDefinition<ComponentProps, Props> =
-  | {
-      Component?: never
-      props: Partial<Props>
-    }
-  | {
-      Component: React.ComponentType<ComponentProps>
-      props?: never
-    }
+export type EmptyStateOverrideDefinition = SectionOverrideDefinition<
+  'EmptyState',
+  {
+    EmptyState: ComponentOverrideDefinition<
+      PropsWithChildren<EmptyStateProps>,
+      EmptyStateProps
+    >
+  }
+>
 
-export type Prettify<T> = {
-  [K in keyof T]: T[K]
-} & {}
+export type HeroOverrideDefinition = SectionOverrideDefinition<
+  'Hero',
+  {
+    Hero: ComponentOverrideDefinition<HeroProps, HeroProps>
+    HeroImage: ComponentOverrideDefinition<HeroImageProps, HeroImageProps>
+    HeroHeader: ComponentOverrideDefinition<HeroHeaderProps, HeroHeaderProps>
+  }
+>
 
-// Taken from https://dev.to/lucianbc/union-type-merging-in-typescript-9al
-type AllKeys<T> = T extends any ? keyof T : never
+export type NavbarOverrideDefinition = SectionOverrideDefinition<
+  'Navbar',
+  {
+    Navbar: ComponentOverrideDefinition<NavbarProps, NavbarProps>
+    NavbarLinks: ComponentOverrideDefinition<NavbarLinksProps, NavbarLinksProps>
+    NavbarLinksList: ComponentOverrideDefinition<
+      NavbarLinksListProps,
+      NavbarLinksListProps
+    >
+    NavbarSlider: ComponentOverrideDefinition<
+      NavbarSliderProps,
+      NavbarSliderProps
+    >
+    NavbarSliderHeader: ComponentOverrideDefinition<
+      NavbarSliderHeaderProps,
+      NavbarSliderHeaderProps
+    >
+    NavbarSliderContent: ComponentOverrideDefinition<
+      NavbarSliderContentProps,
+      NavbarSliderContentProps
+    >
+    NavbarSliderFooter: ComponentOverrideDefinition<
+      NavbarSliderFooterProps,
+      NavbarSliderFooterProps
+    >
+    NavbarHeader: ComponentOverrideDefinition<
+      NavbarHeaderProps,
+      NavbarHeaderProps
+    >
+    NavbarRow: ComponentOverrideDefinition<NavbarRowProps, NavbarRowProps>
+    NavbarButtons: ComponentOverrideDefinition<
+      NavbarButtonsProps,
+      NavbarButtonsProps
+    >
+    IconButton: ComponentOverrideDefinition<
+      IconButtonProps,
+      Omit<IconButtonProps, 'onClick'>
+    >
+  }
+>
 
-type PickType<T, K extends AllKeys<T>> = T extends { [k in K]?: any }
-  ? T[K]
-  : undefined
+export type NewsletterOverrideDefinition = SectionOverrideDefinition<
+  'Newsletter',
+  {
+    ToastIconSuccess: ComponentOverrideDefinition<IconProps, IconProps>
+    ToastIconError: ComponentOverrideDefinition<IconProps, IconProps>
+    HeaderIcon: ComponentOverrideDefinition<IconProps, IconProps>
+    InputFieldName: ComponentOverrideDefinition<
+      InputFieldProps,
+      Omit<InputFieldProps, 'inputRef'>
+    >
+    InputFieldEmail: ComponentOverrideDefinition<
+      InputFieldProps,
+      Omit<InputFieldProps, 'inputRef'>
+    >
+    Button: ComponentOverrideDefinition<ButtonProps, ButtonProps>
+  }
+>
 
-type Merge<T> = {
-  [k in AllKeys<T>]: PickType<T, k>;
-};
+export type ProductDetailsOverrideDefinition = SectionOverrideDefinition<
+  'ProductDetails',
+  {
+    ProductTitle: ComponentOverrideDefinition<
+      ProductTitleProps,
+      ProductTitleProps
+    >
+    DiscountBadge: ComponentOverrideDefinition<
+      DiscountBadgeProps,
+      Omit<DiscountBadgeProps, 'listPrice' | 'spotPrice'>
+    >
+    BuyButton: ComponentOverrideDefinition<ButtonProps, ButtonProps>
+    Icon: ComponentOverrideDefinition<IconProps, IconProps>
+    Price: ComponentOverrideDefinition<
+      PriceProps,
+      Omit<PriceProps, 'value' | 'data-value'>
+    >
+    QuantitySelector: ComponentOverrideDefinition<
+      QuantitySelectorProps,
+      Omit<QuantitySelectorProps, 'onChange'>
+    >
+    SkuSelector: ComponentOverrideDefinition<SkuSelectorProps, SkuSelectorProps>
+    ShippingSimulation: ComponentOverrideDefinition<
+      ShippingSimulationProps,
+      ShippingSimulationProps
+    >
+    ImageGallery: ComponentOverrideDefinition<
+      ImageGalleryProps,
+      ImageGalleryProps
+    >
+    ImageZoom: ComponentOverrideDefinition<ImageZoomProps, ImageZoomProps>
+    __experimentalImageGalleryImage: ComponentOverrideDefinition<any, any>
+    __experimentalImageGallery: ComponentOverrideDefinition<any, any>
+    __experimentalShippingSimulation: ComponentOverrideDefinition<any, any>
+  }
+>
+
+export type ProductGalleryOverrideDefinition = SectionOverrideDefinition<
+  'ProductGallery',
+  {
+    Button: ComponentOverrideDefinition<
+      ButtonProps,
+      Omit<ButtonProps, 'onClick'>
+    >
+    FilterIcon: ComponentOverrideDefinition<IconProps, IconProps>
+    PrevIcon: ComponentOverrideDefinition<IconProps, IconProps>
+    ResultsCountSkeleton: ComponentOverrideDefinition<
+      SkeletonProps,
+      Omit<SkeletonProps, 'loading'>
+    >
+    SortSkeleton: ComponentOverrideDefinition<
+      SkeletonProps,
+      Omit<SkeletonProps, 'loading'>
+    >
+    FilterButtonSkeleton: ComponentOverrideDefinition<
+      SkeletonProps,
+      Omit<SkeletonProps, 'loading'>
+    >
+    LinkButtonPrev: ComponentOverrideDefinition<
+      LinkButtonProps,
+      Omit<LinkButtonProps, 'onClick' | 'href'>
+    >
+    LinkButtonNext: ComponentOverrideDefinition<
+      LinkButtonProps,
+      Omit<LinkButtonProps, 'onClick' | 'href'>
+    >
+    __experimentalFilterDesktop: ComponentOverrideDefinition<any, any>
+    __experimentalFilterSlider: ComponentOverrideDefinition<any, any>
+    __experimentalProductCard: ComponentOverrideDefinition<any, any>
+  }
+>
+
+export type ProductShelfOverrideDefinition = SectionOverrideDefinition<
+  'ProductShelf',
+  {
+    ProductShelf: ComponentOverrideDefinition<
+      ProductShelfProps,
+      ProductShelfProps
+    >
+    __experimentalCarousel: ComponentOverrideDefinition<any, any>
+    __experimentalProductCard: ComponentOverrideDefinition<
+      any,
+      Omit<any, 'key' | 'product' | 'index'>
+    >
+  }
+>
+
+export type RegionBarOverrideDefinition = SectionOverrideDefinition<
+  'RegionBar',
+  {
+    RegionBar: ComponentOverrideDefinition<
+      RegionBarProps,
+      Omit<RegionBarProps, 'onButtonClick' | 'postalCode'>
+    >
+    LocationIcon: ComponentOverrideDefinition<IconProps, IconProps>
+    ButtonIcon: ComponentOverrideDefinition<IconProps, IconProps>
+  }
+>
